Scheduling PoC using DynamoDB and Quartz scheduler
The idea is to use the atomicity of DynamoDB's conditional write to
provide load balancing and ensure that each schedule is only processed
once. Multiple instances of the service may try to orchestrate a given
comm, but Dynamo ensures that only one instance will successfully
execute the conditional update needed to take ownership of that comm. In
other words we use Dynamo like a distributed lock.

A scheduled comm starts in Pending state. When its scheduled delivery
time comes around, it moves into Orchestrating state. After
orchestration has executed, it moves to either Complete or Failed. If a
cancellation event arrives while the comm is still in Pending state, it
moves to Cancelled.

```
Pending -> Orchestrating -> Complete
                         -> Failed
        -> Cancelled
```

Each comm is stored as an item in DynamoDB. Even comms with no schedule
(i.e. those that should be sent immediately) are treated as scheduled
comms with delivery date == now.

When a Triggered event arrives:

(status: not implemented yet, but should be simple)

1. Create a Dynamo item with Pending status
2. Create a job and register it with the Quartz scheduler

When a job executes:

(status: proof of concept implemented)

1. Conditionally update the Dynamo item to set status = Orchestrating
   and expiry = now + 5 minutes, iff status == Pending OR
   (status == Orchestrating AND expiry < now)
2. If the update succeeded:
    1. Do the orchestration and send a Kafka event
    2. Update the Dynamo record to set status = Complete (or Failed)
   If the update failed, it means either the comm was cancelled or
   another instance of the comm beat us to it, so there is nothing to do

When an instance starts up:

(status: proof of concept implemented)

1. Load all Dynamo items with status == Pending OR
   (status == Orchestrating AND expiry < now)
2. For each item, create a job and register it with the scheduler
3. Unfortunately there is a race condition here, so do the same thing a
   few mins later to catch any items that we missed

Assumptions:

* Orchestration of a comm always takes < 5 minutes
* No clock skew

Risks/known issues:

* If an orchestration times out, nobody will pick it up until a new
  instance is started
* If orchestration succeeds (i.e. sends a Kafka event) but it then
  fails to update the Dynamo item, the comm will be orchestrated again
* Potentially large Dynamo read load at startup